C++ Astrology program 1
Write an astrology program. The user types in a birthday, and the program responds with the sign and horoscope for that birthday. The month may be entered as a number from 1 to 12. Then enhance your program so that if the
birthday is only one or two days away from an adjacent sign, the program announces that the birthday is on a “cusp” and outputs the horoscope for that nearest adjacent sign. This program will have a long multiway branch. Make up a
horoscope for each sign. The horoscope signs and dates are:
Aries March 21–April 19
Taurus April 20–May 20
Gemini May 21–June 21
Cancer June 22–July 22
Leo July 23–August 22
Virgo August 23–September 22
Libra September 23–October 22
Scorpio October 23–November 21
Sagittarius November 22–December 21
Capricorn December 22–January 19
Aquarius January 20–February 18
Pisces February 19–March 20